+++ demo/src/main/java/org/ankarajug/testinfected/service/CalculatorServiceImpl.java
@@ -22,8 +22,8 @@ public class CalculatorServiceImpl implements CalculatorService {
 
     @Override
     @Transactional
-    public CalculationResult add(double a, double b) {
-        CalculationResult result = new CalculationResultBuilder().action(Action.ADD).param1(a).param2(b).result(a + b).build();
+    public CalculationResult add(Integer a, Integer b) {
+        CalculationResult result = new CalculationResultBuilder().action(Action.ADD).param1(a).param2(b).result(Double.valueOf(a + b)).build();
         dao.save(result);
 
         return result;
@@ -31,8 +31,8 @@ public class CalculatorServiceImpl implements CalculatorService {
 
     @Override
     @Transactional
-    public CalculationResult subtract(double a, double b) {
-        CalculationResult result = new CalculationResultBuilder().action(Action.SUBTRACT).param1(a).param2(b).result(a - b).build();
+    public CalculationResult subtract(Integer a, Integer b) {
+        CalculationResult result = new CalculationResultBuilder().action(Action.SUBTRACT).param1(a).param2(b).result(Double.valueOf(a - b)).build();
         dao.save(result);
 
         return result;
@@ -40,8 +40,8 @@ public class CalculatorServiceImpl implements CalculatorService {
 
     @Override
     @Transactional
-    public CalculationResult multiply(double a, double b) {
-        CalculationResult result = new CalculationResultBuilder().action(Action.MULTIPLY).param1(a).param2(b).result(a * b).build();
+    public CalculationResult multiply(Integer a, Integer b) {
+        CalculationResult result = new CalculationResultBuilder().action(Action.MULTIPLY).param1(a).param2(b).result(Double.valueOf(a * b)).build();
         dao.save(result);
 
         return result;
@@ -49,8 +49,8 @@ public class CalculatorServiceImpl implements CalculatorService {
 
     @Override
     @Transactional
-    public CalculationResult divide(double a, double b) {
-        CalculationResult result = new CalculationResultBuilder().action(Action.DIVIDE).param1(a).param2(b).result(a / b).build();
+    public CalculationResult divide(Integer a, Integer b) {
+        CalculationResult result = new CalculationResultBuilder().action(Action.DIVIDE).param1(a).param2(b).result((double)a / (double)b).build();
         dao.save(result);
 
         return result;
